requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.16.6
profanity-check==1.0.3
PyGithub==1.45
twilio==6.0.0
scikit-learn==0.20.4
pyap
geopy
folium